requirements.txt
streamlit
streamlit-oauth
langchain
langchain-groq
google-api-python-client
python-dotenv